attached_assets/Pasted-Error-Element-type-is-invalid-expected-a-string-for-built-in-components-or-a-class-function-for-1735489471885.txt

Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.

Check the render method of `UploadPage`.
    at createFiberFromTypeAndProps (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:13366:28)
    at createFiberFromElement (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:13380:14)
    at reconcileChildFibersImpl (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:7012:31)
    at eval (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:7189:33)
    at reconcileChildren (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:7752:13)
    at beginWork (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:9957:13)
    at runWithFiberInDEV (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:544:16)
    at performUnitOfWork (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:15064:22)
    at workLoopSync (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:14888:41)
    at renderRootSync (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:14868:11)
    at performWorkOnRoot (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:14394:44)
    at performWorkOnRootViaSchedulerTask (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:15955:7)
    at MessagePort.performWorkUntilDeadline (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/scheduler/cjs/scheduler.development.js:44:48)
